Get requests (#7)
* Define getTickets method

This also defines a very minimal Request interface. This only contains
the minimum needed for us to complete some work around showing the
latest updated ticket. It can be expanded as need be.

Impelmentations to follow.

* Add type for RequestStatus

This makes the choice available for statuses more explicit and easier to
understand on first glance.

* Android implements getRequests method

* Add getRequests methods for iOS

* Use epoch time for ticket updatedAt

* Use lowercase ticket status in android

* remove unused date formatter code

Co-authored-by: Adam Saeid <adam@residently.com>